feat: add background handler (#39)
This provides a working background script with an rpc provider and client to communicate with background script.
Worth to consider that all calls from different sections for making changes in chrome storage and utilizing other core extension APIs  will go through the provided internal RPC that is setup to invoke/handle those calls in the background scripts. This solves the state initiation problem and helps with properly notifying the dapps regarding relevant state changes.

This is the pattern that we are following for adding RPC calls:

1-  The rpc call signatures are defined as ProviderRpc type. The signature defines the ```rpc call message type``` as well as expected ```Request``` and ```Response``` format.
 Currently [this implements AccountProviderRpc](https://github.com/paritytech/polkadot-lite-extension/blob/4a944b6592ef69514c8366b7b51e5bba44af5067/packages/core/provider/types.ts#L117) but we might have more for signer and other type of providers as we add more functionality.
e.g.  the rpc call signature for loading PJS accounts is defined as following 
https://github.com/paritytech/polkadot-lite-extension/blob/4a944b6592ef69514c8366b7b51e5bba44af5067/packages/core/provider/types.ts#L159

2- After the call signature is defined, a proper handler for the rpc call needs to be implemented in the [ProviderMessageHandler](https://github.com/paritytech/polkadot-lite-extension/blob/ea125bbb9a2da9a58dbf174e9a731ebd522fd906/packages/core/provider/index.ts#L53) to handle the call.
e.g. this is were an example of how  account import handler is implemented:
https://github.com/paritytech/polkadot-lite-extension/blob/4a944b6592ef69514c8366b7b51e5bba44af5067/packages/core/provider/handler.ts#L50

3- A method needs to be added to the rpcClient to let clients send the correct rpc message according to defined RPC signature.
The rpc client method only needs to send the correct message based on the defined rpc signature in step 1 and if needed transform the returned rpc response to prepare it  to be used in other components.
e.g. this is an example method in rpc client for importing PJS:
https://github.com/paritytech/polkadot-lite-extension/blob/4a944b6592ef69514c8366b7b51e5bba44af5067/apps/extension/src/rpc-client/index.ts#L68

4- The [RpcClient](https://github.com/paritytech/polkadot-lite-extension/blob/ea125bbb9a2da9a58dbf174e9a731ebd522fd906/packages/core/client/index.ts#L20) can be used by other internal and external components and pages to communicate with the background script and extension internal Apis (storage, notifications, ...)
e.g. this is an example how the rpc client is used in import component.
https://github.com/paritytech/polkadot-lite-extension/blob/ea125bbb9a2da9a58dbf174e9a731ebd522fd906/packages/import/index.tsx#L42